split category management page
